+++ CHANGELOG.rst
@@ -5,6 +5,8 @@ Current release candidate
 -------------------------
 
 * Fixed bug with deleting models from tasks
+* Support working with model meta without requiring installation of all model dependencies
+
 
 0.3.3 (2020-01-10)
 ------------------